Move `maker_negotiator.ts` to `maker/`
